Avoid truncation for inputs close to limit
Previous truncation code always truncated and added a tail once the
processed input width exceeded the limit minus the tail width. For
example, the input "foobar" truncated to width 6 with a tail "." would
come out as "fooba.".

Now, when the rune scanning loop passes into the danger zone where
truncation might be required, but before the truncation limit is
breached, it buffers runes instead of writing them out. If the
truncation limit is exceeded, then the tail is written as usual. If it
ends up not exceeded, then the buffered runes are written out, avoiding
unnecessary truncation.