fix wordwrap of hyphens at limit
If the word-thus-far is exactly at the limit and `wordwrap.Write` sees a hyphen, it writes the word to the buffer *and the hyphen*, exceeding the limit. This change fixes that behavior by instead adding the byphen to the word "as normal", and *then* checking to see if the word will fit within the limit. If it does, only then will it flush the word to the buffer.

Fix #66